Compute the same bits on arm64 as on amd64
The goldens did not reproduce on linux/arm64, the console's architecture.
There the compiler fuses a product that feeds an addition or subtraction
into one FMADD/FMSUB instruction with a single rounding, where amd64 rounds
twice; about a quarter of all inputs differ in the last bit, and inlining
carries the fusion across statements and function boundaries. gmath/trig.go
and Mat4.MulPoint already guarded against it; the rest of the engine did not.

Every product that feeds + or - in gmath, gfx/soft, sprite, scene, sim,
asset/texture, asset/model, inspect and the headless sheets is now wrapped
in an explicit conversion to its own type, which the Go specification makes
a rounding point the compiler may not fuse across. On amd64 the conversions
change nothing: go test ./... passes with no golden touched, and the
rasterizer benchmark is unchanged within noise (0 allocs/op). Under
qemu-aarch64 the whole suite now passes against the amd64 goldens, including
the four scenario hashes and contact sheets, the model sheets and inspect.

internal/fused keeps it that way: it reads a linux/arm64 ELF, finds the
fused instructions by their encoding and maps them to source lines through
the Go line table. TestEngineHasNoFusedMultiplyAdd builds the veduta tool
and the template game for arm64 and fails on any engine line that fused
(133 lines before this change, 0 after); a control test proves the scanner
finds a known fusion, so the gate cannot pass by finding nothing.

Two read-only audits looked for other architecture differences. The math
routines that are not exact everywhere were used twice, both in inspect:
math.Cbrt now counts cells in integers and math.Log ranks density ratios as
max(r, 1/r). Out-of-range float to int conversions differ (NaN and values
past 2^31 or 2^63 saturate on arm64); the debug line drawer, the label
placement of scene sheets and the simulate tile height now refuse or clamp
those values in float first, with the result amd64 already produced.

+++ asset/model/geom.go
@@ -156,11 +156,11 @@ func firstCrossing(pts []dvec2) (int, int, bool) {
 			c, d := pts[j], pts[(j+1)%n]
 			switch {
 			case j == i+1: // edges meet at b == c: bad only if d folds back onto a-b
-				if cross2(a, b, d) == 0 && (d[0]-b[0])*(a[0]-b[0])+(d[1]-b[1])*(a[1]-b[1]) > 0 {
+				if cross2(a, b, d) == 0 && float64((d[0]-b[0])*(a[0]-b[0]))+float64((d[1]-b[1])*(a[1]-b[1])) > 0 {
 					return i, j, true
 				}
 			case i == 0 && j == n-1: // edges meet at a == d
-				if cross2(c, a, b) == 0 && (c[0]-a[0])*(b[0]-a[0])+(c[1]-a[1])*(b[1]-a[1]) > 0 {
+				if cross2(c, a, b) == 0 && float64((c[0]-a[0])*(b[0]-a[0]))+float64((c[1]-a[1])*(b[1]-a[1])) > 0 {
 					return i, j, true
 				}
 			default:
